fix bug/misleading output in GLG4Scint
Despite what it printed, GLG4Scint was modified to only produce
scintillation photons if the LIGHT_YIELD parameter was set for the
material.

This restores the default behavior: calculating the light yield from the
spectrum normalization and overriding with LIGHT_YIELD if it is
specified.

Thanks to @twongjirad for pointing this out!