implement bnl wls model
This is the bulk wavelength shifter model written by L. Bignell at BNL
and first implemented in RAT-PAC in #20.

I've modified it to work with the current RAT-PAC Geant4 10 implementation,
and made some other small changes:

* Changed name from "Alt" to "BNL"
* Moved hardcoded data file name to RATDB
* Moved file loading to its own class, to avoid G4/ROOT name conflicts,
  and keep compiling with -Werror

At this point, this compiles and runs but has not been tested; submitted
for initial review but perhaps await some validation before merging.