add the decay0 generator
Add the Decay0 Generator for Initial Kinematics in Alpha, Beta and
Double Beta Decays.

The Decay0 code was originally written by V. Tretyak (in Fortran) [1]
and later ported to C++ by members of the SNO+ Collaboration. This
version is the SNO+ version, released with permission of Tretyak and
the SNO+ Board. For more information, see COPYING.decay0 and
src/gen/Decay0.hh.